Split main_LinuxX11.cpp from main_Linux.cpp
